// lib/types.generated.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export type AboutUsDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<AboutUsDocumentData>, "about_us", Lang>;
// ...
export type AuthorsDocument<Lang extends string = string> = prismic.PrismicDocumentWithoutUID<Simplify<AuthorsDocumentData>, "authors", Lang>;
// ...
export type BlogDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<BlogDocumentData>, "blog", Lang>;
// ...
export type BlogPostDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<BlogPostDocumentData>, "blog_post", Lang>;
// ...
export type BondPriceCalculatorDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<BondPriceCalculatorDocumentData>, "bond_price_calculator", Lang>;
// ...
export interface ChildPortfolioDocumentDataBodyFaqSlicePrimary {
	/**
	 * Intro field in *Child Portfolio → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: child_portfolio.body[].faq.primary.intro
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	intro: prismic.RichTextField;
	
	/**
	 * More button label field in *Child Portfolio → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: child_portfolio.body[].faq.primary.more_button_label
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	more_button_label: prismic.KeyTextField;
}

/**
 * Item content in *Child Portfolio → Slice zone → FAQ → Items*
 */
export interface ChildPortfolioDocumentDataBodyFaqSliceItem {
	/**
	 * FAQ Item field in *Child Portfolio → Slice zone → FAQ → Items*
	 *
	 * - **Field Type**: Content Relationship
	 * - **Placeholder**: *None*
	 * - **API ID Path**: child_portfolio.body[].faq.items.faq_item
	 * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
	 */
	faq_item: prismic.ContentRelationshipField<"faq_item">;
}

/**
 * Slice for *Child Portfolio → Slice zone*
 */
export type ChildPortfolioDocumentDataBodyFaqSlice = prismic.Slice<"faq", Simplify<ChildPortfolioDocumentDataBodyFaqSlicePrimary>, Simplify<ChildPortfolioDocumentDataBodyFaqSliceItem>>
// ...
export type ChildPortfolioDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<ChildPortfolioDocumentData>, "child_portfolio", Lang>;
// ...
export type CompoundInterestCalculatorDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<CompoundInterestCalculatorDocumentData>, "compound_interest_calculator", Lang>;
// ...
export type ContactDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<ContactDocumentData>, "contact", Lang>;
// ...
export type CultureDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<CultureDocumentData>, "culture", Lang>;
// ...
export type FaqDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<FaqDocumentData>, "faq", Lang>;
// ...
type FaqCategoryDocumentDataBodySlice = FaqCategoryDocumentDataBodySubcategorySlice

/**
 * Content for FAQ Category documents
 */
interface FaqCategoryDocumentData {
	/**
	 * Title field in *FAQ Category*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.title
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	title: prismic.RichTextField;
	
	/**
	 * Description field in *FAQ Category*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.description
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	description: prismic.RichTextField;
	
	/**
	 * Icon field in *FAQ Category*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.icon
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#image
	 */
	icon: prismic.ImageField<never>;
	
	/**
	 * Slice zone field in *FAQ Category*
	 *
	 * - **Field Type**: Slice Zone
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.body[]
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#slices
	 */
	body: prismic.SliceZone<FaqCategoryDocumentDataBodySlice>;/**
	 * SEO Title field in *FAQ Category*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.seo_title
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	seo_title: prismic.KeyTextField;
	
	/**
	 * SEO Description field in *FAQ Category*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.seo_description
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	seo_description: prismic.KeyTextField;
	
	/**
	 * Social Media Image field in *FAQ Category*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: faq_category.social_media_image
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#image
	 */
	social_media_image: prismic.ImageField<never>;
}

/**
 * FAQ Category document from Prismic
 *
 * - **API ID**: `faq_category`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type FaqCategoryDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<FaqCategoryDocumentData>, "faq_category", Lang>;
// ...
export type FaqItemDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<FaqItemDocumentData>, "faq_item", Lang>;
// ...
export type FaqOverviewDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<FaqOverviewDocumentData>, "faq_overview", Lang>;
// ...
export type FooterNavigationDocument<Lang extends string = string> = prismic.PrismicDocumentWithoutUID<Simplify<FooterNavigationDocumentData>, "footer_navigation", Lang>;
// ...
export interface HomepageDocumentDataBodyBlogPostsSlicePrimary {
	/**
	 * Amount of posts to show field in *Homepage → Slice zone → Blog Posts → Primary*
	 *
	 * - **Field Type**: Number
	 * - **Placeholder**: *None*
	 * - **API ID Path**: homepage.body[].blog_posts.primary.posts_amount
	 * - **Documentation**: https://prismic.io/docs/field#number
	 */
	posts_amount: prismic.NumberField;
}

/**
 * Slice for *Homepage → Slice zone*
 */
export type HomepageDocumentDataBodyBlogPostsSlice = prismic.Slice<"blog_posts", Simplify<HomepageDocumentDataBodyBlogPostsSlicePrimary>, never>
// ...
export type HomepageDocument<Lang extends string = string> = prismic.PrismicDocumentWithoutUID<Simplify<HomepageDocumentData>, "homepage", Lang>;
// ...
export type HowItWorksDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<HowItWorksDocumentData>, "how_it_works", Lang>;
// ...
export type JobsDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<JobsDocumentData>, "jobs", Lang>;
// ...
export type KnowledgeDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<KnowledgeDocumentData>, "knowledge", Lang>;
// ...
export type KnowledgeNavigationDocument<Lang extends string = string> = prismic.PrismicDocumentWithoutUID<Simplify<KnowledgeNavigationDocumentData>, "knowledge_navigation", Lang>;
// ...
export type MainNavigationDocument<Lang extends string = string> = prismic.PrismicDocumentWithoutUID<Simplify<MainNavigationDocumentData>, "main_navigation", Lang>;
// ...
export type MaintenanceDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<MaintenanceDocumentData>, "maintenance", Lang>;
// ...
export type PagesDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<PagesDocumentData>, "pages", Lang>;
// ...
export interface Pillar3aDocumentDataBodyFaqSlicePrimary {
	/**
	 * Intro field in *Pillar 3a → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pillar_3a.body[].faq.primary.intro
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	intro: prismic.RichTextField;
	
	/**
	 * More button label field in *Pillar 3a → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pillar_3a.body[].faq.primary.more_button_label
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	more_button_label: prismic.KeyTextField;
}

/**
 * Item content in *Pillar 3a → Slice zone → FAQ → Items*
 */
export interface Pillar3aDocumentDataBodyFaqSliceItem {
	/**
	 * FAQ Item field in *Pillar 3a → Slice zone → FAQ → Items*
	 *
	 * - **Field Type**: Content Relationship
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pillar_3a.body[].faq.items.faq_item
	 * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
	 */
	faq_item: prismic.ContentRelationshipField<"faq_item">;
}

/**
 * Slice for *Pillar 3a → Slice zone*
 */
export type Pillar3aDocumentDataBodyFaqSlice = prismic.Slice<"faq", Simplify<Pillar3aDocumentDataBodyFaqSlicePrimary>, Simplify<Pillar3aDocumentDataBodyFaqSliceItem>>
// ...
export type Pillar3aDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<Pillar3aDocumentData>, "pillar_3a", Lang>;
// ...
export type Pillar3aTaxSavingCalculatorDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<Pillar3aTaxSavingCalculatorDocumentData>, "pillar_3a_tax_saving_calculator", Lang>;
// ...
export type PressDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<PressDocumentData>, "press", Lang>;
// ...
export type PressReleaseDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<PressReleaseDocumentData>, "press_release", Lang>;

/**
 * Slice for *Pricing → Slice zone*
 */
export type PricingDocumentDataBodyReadyToInvestSlice = prismic.Slice<"ready_to_invest", Record<string, never>, never>

/**
 * Primary content in *Pricing → Slice zone → FAQ → Primary*
 */
export interface PricingDocumentDataBodyFaqSlicePrimary {
	/**
	 * Intro field in *Pricing → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pricing.body[].faq.primary.intro
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	intro: prismic.RichTextField;
	
	/**
	 * More text field in *Pricing → Slice zone → FAQ → Primary*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pricing.body[].faq.primary.more_text
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	more_text: prismic.RichTextField;
}

/**
 * Item content in *Pricing → Slice zone → FAQ → Items*
 */
export interface PricingDocumentDataBodyFaqSliceItem {
	/**
	 * FAQ Item field in *Pricing → Slice zone → FAQ → Items*
	 *
	 * - **Field Type**: Content Relationship
	 * - **Placeholder**: *None*
	 * - **API ID Path**: pricing.body[].faq.items.faq_item
	 * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
	 */
	faq_item: prismic.ContentRelationshipField<"faq_item">;
}

/**
 * Slice for *Pricing → Slice zone*
 */
export type PricingDocumentDataBodyFaqSlice = prismic.Slice<"faq", Simplify<PricingDocumentDataBodyFaqSlicePrimary>, Simplify<PricingDocumentDataBodyFaqSliceItem>>
// ...
export type PricingDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<PricingDocumentData>, "pricing", Lang>;
// ...
export type ReferralInvitationDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<ReferralInvitationDocumentData>, "referral_invitation", Lang>;
// ...
export type SustainableInvestmentDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<SustainableInvestmentDocumentData>, "sustainable_investment", Lang>;
// ...
export type WhoIsItForDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<WhoIsItForDocumentData>, "who_is_it_for", Lang>;
// ...
export type WhyTrueWealthDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<WhyTrueWealthDocumentData>, "why_true_wealth", Lang>;

export type AllDocumentTypes = AboutUsDocument | AuthorsDocument | BlogDocument | BlogPostDocument | BondPriceCalculatorDocument | ChildPortfolioDocument | CompoundInterestCalculatorDocument | ContactDocument | CultureDocument | FaqDocument | FaqCategoryDocument | FaqItemDocument | FaqOverviewDocument | FooterNavigationDocument | HomepageDocument | HowItWorksDocument | JobsDocument | KnowledgeDocument | KnowledgeNavigationDocument | MainNavigationDocument | MaintenanceDocument | PagesDocument | Pillar3aDocument | Pillar3aTaxSavingCalculatorDocument | PressDocument | PressReleaseDocument | PricingDocument | ReferralInvitationDocument | SustainableInvestmentDocument | WhoIsItForDocument | WhyTrueWealthDocument;